Update [#0022]
Update [#0022]

#Important Update Warning#
#      Beta Version      #

1. Add result section.
2. Add excel export section.
3. Optimize the display of each form.

Summary @ 2015-08-25 05:26:26 UTC+8

+++ admin/ListViewBooksSorter.cs
@@ -37,12 +37,12 @@ namespace PublishServer
             itemY = y as System.Windows.Forms.ListViewItem;
             string xText = itemX.SubItems[ColumnToSort].Text;
             string yText = itemY.SubItems[ColumnToSort].Text;
-            int xNum, yNum;
+            double xNum, yNum;
             switch (ColumnToSort)
             {
                 case 0: case 7:
-                    xNum = int.Parse(xText);
-                    yNum = int.Parse(yText);
+                    xNum = double.Parse(xText);
+                    yNum = double.Parse(yText);
                     cmpResult = xNum.CompareTo(yNum);
                     break;
                 case 10:
